Struct dashu_int::modular::Modulo

pub struct Modulo<'a>(_);

Modular arithmetic.

Examples

let ring = ModuloRing::new(UBig::from(10000u32));
let x = ring.convert(12345);
let y = ring.convert(55443);
assert_eq!((x - y).residue(), UBig::from(6902u32));

Implementations

impl Modulo<'_>

pub fn residue(&self) -> UBig

Get the residue in range 0..n in an n-element ring.

Examples

let ring = ModuloRing::new(UBig::from(100u8));
let x = ring.convert(-1234);
assert_eq!(x.residue(), UBig::from(66u8));

impl<'a> Modulo<'a>

pub fn inv(&self) -> Option<Modulo<'a>>

Multiplicative inverse.

Examples

// A Mersenne prime.
let p = UBig::from(2u8).pow(127) - UBig::ONE;
let ring = ModuloRing::new(p.clone());
// Fermat's little theorem: a^(p-2) = a^-1 (mod p)
let a = ring.convert(123);
let ainv = a.clone().inv().unwrap();
assert_eq!(ainv, a.pow(&(p - UBig::from(2u8))));
assert_eq!((a * ainv).residue(), UBig::ONE);

impl<'a> Modulo<'a>

pub fn pow(&self, exp: &UBig) -> Modulo<'a>

Exponentiation.

If you want use a negative exponent, you can first use inv() to convert the base to its inverse, and then call this method.

Examples

// A Mersenne prime.
let p = UBig::from(2u8).pow(607) - UBig::ONE;
let ring = ModuloRing::new(p.clone());
// Fermat's little theorem: a^(p-1) = 1 (mod p)
let a = ring.convert(123);
assert_eq!(a.pow(&(p - UBig::ONE)), ring.convert(1));