Trait dashu_int::ops::BitTest

pub trait BitTest {
    fn bit_len(&self) -> usize;
    fn bit(&self, n: usize) -> bool;
    fn trailing_zeros(&self) -> Option<usize>;
}

Common bit operations for integers

Required Methods

fn bit_len(&self) -> usize

Get the minimum required number of bits to represent this integer

fn bit(&self, n: usize) -> bool

Get the n-th bit of the integer

fn trailing_zeros(&self) -> Option<usize>

Get the number of trailing zeros in the integer