Trait dashu_int::ops::Gcd

pub trait Gcd<Rhs = Self> {
    type Output;

    // Required method
    fn gcd(self, rhs: Rhs) -> Self::Output;
}

Compute the greatest common divisor.

For negative integers, the common divisor is still kept positive.

Examples

use dashu_base::Gcd;
assert_eq!(12u8.gcd(10u8), 2);

Panics

Panics if both operands are zeros

Required Associated Types

type Output

Required Methods

fn gcd(self, rhs: Rhs) -> Self::Output

Compute the greatest common divisor between the two operands.

Panics if both operands are zeros